+++ src/client/cypress/e2e/admin/aboutSettings.cy.js
@@ -3,7 +3,7 @@ import { goToRouteAndAcceptTerms } from "../helpers/testHelpers.js";
 
 describe("Admin about page tests", () => {
   it("shows version information linking the corresponding release on GitHub.", () => {
-    goToRouteAndAcceptTerms("/setting/about");
+    goToRouteAndAcceptTerms("/setting#about");
 
     cy.get('[data-cy="version"]')
       .should("contain", "0.0.99+dev")
@@ -12,15 +12,14 @@ describe("Admin about page tests", () => {
 
   it("shows license information (with fixtures)", () => {
     cy.intercept("/license.json", license);
-    goToRouteAndAcceptTerms("/setting/about");
-
+    goToRouteAndAcceptTerms("/setting#about");
     cy.get('[data-cy^="credits-"]').should("have.length", 2);
     cy.get('[data-cy="credits-example-js@0.0.999"]').should("contain", "example-js (Version 0.0.999)");
     cy.get('[data-cy="credits-example-react@0.0.111"]').should("contain", "example-react (Version 0.0.111)");
   });
 
   it("shows license information (without fixtures)", () => {
-    goToRouteAndAcceptTerms("/setting/about");
+    goToRouteAndAcceptTerms("/setting#about");
     cy.get('[data-cy^="credits-"]').should("have.length.above", 0);
   });
 });