+++ src/client/cypress/e2e/settings/mapSettings.cy.js
@@ -1,4 +1,4 @@
-import { goToRouteAndAcceptTerms, returnToOverview } from "../helpers/testHelpers.js";
+import { getElementByDataCy, goToRouteAndAcceptTerms, returnToOverview } from "../helpers/testHelpers.js";
 
 describe("map settings", () => {
   it("Adds wms and wmts to user maps", () => {
@@ -7,6 +7,7 @@ describe("map settings", () => {
     const wmsName = "Army logistics centres (ALC)";
     const wmtsName = "Wetness potential agricultural land";
 
+    getElementByDataCy("general-tab").click();
     cy.contains("Map").click();
     // Add WMS
     cy.get('[data-cy="load-layers-button"]').click();
@@ -38,7 +39,8 @@ describe("map settings", () => {
     goToRouteAndAcceptTerms("/");
 
     // Remove layers
-    cy.get('[data-cy="settings-button"]').click();
+    getElementByDataCy("settings-button").click();
+    getElementByDataCy("general-tab").click();
     cy.contains("Map").click();
     cy.wait(1000);
     cy.get('[data-cy="maps-for-user-box"]').contains(wmtsName);