+++ src/client/src/api-lib/ReduxStateInterfaces.ts
@@ -42,6 +42,7 @@ export interface UserData {
   id: number;
   name: string;
   username: string;
+  admin: boolean;
 }
 
 export interface Workgroup {